requirements.txt
django>=4.2.0,<5.0.0
django-tenants>=3.5.0
django-ninja>=1.0.0
psycopg2-binary>=2.9.0 